Backend/requirements.txt
fastapi 
bcrypt == 4.0.1
uvicorn
alembic
sqlalchemy
httpx
passlib
python-jose
secure-smtplib==0.1.1
python-dotenv==1.0.1
psycopg2-binary
requests
watchgod